backend/pyproject.toml
[tool.poetry]
name = "buildai-backend"
version = "0.1.0"
description = "BuildAI Blueprint Analysis API"
authors = ["BuildAI"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
supabase = "^2.3.0"
httpx = "^0.26.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
celery = {extras = ["redis"], version = "^5.3.6"}
redis = "^5.0.1"
boto3 = "^1.34.11"
pillow = "^10.2.0"
opencv-python-headless = "^4.9.0.80"
pytesseract = "^0.3.10"
anthropic = "^0.16.0"
ultralytics = "^8.1.0"
stripe = "^7.8.0"
python-dotenv = "^1.0.0"
reportlab = "^4.0.8"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
ruff = "^0.1.14"
